Revamp hash interface to provide fastpath for byte slices
When dealing with a byte slice, we can dispatch much more quickly
into the Xxh3 fastpath. This results in a 30% faster path when using
Xxh3 (small regression looks like for the std variant).

When using this interface Xxh3 is always faster than std and the
performance differential increases the larger each individual key is.